backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.12
httpx==0.27.2
Pillow==10.4.0
rembg[gpu]>=2.0.62
litellm>=1.57.0
python-dotenv==1.0.1
aiofiles==24.1.0
pytest>=8.0
pytest-asyncio>=0.23
json-repair>=0.30.0
